refactored js files for better reading
