+++ app/views/tweets/feed.html.erb
@@ -27,3 +27,5 @@
     </div>
   </div>
 </div>
+
+    <%= javascript_pack_tag 'feed' %>
