+++ app/views/tweets/index.html.erb
@@ -31,4 +31,5 @@
 </div>
 
 
+    <%= javascript_pack_tag 'index' %>
