+++ app/views/users/show.html.erb
@@ -19,3 +19,4 @@
 <div id="following_users" style="display: none;"> <%= current_user.following_users %> </div>
 
 
+    <%= javascript_pack_tag 'show' %>
